+++ src/components/frame/header/Header.tsx
@@ -1,8 +1,8 @@
 import CartButton from "@/components/cart/cartButton/CartButton";
 import ToFavoriteButton from "@/components/favorite/toFavoriteButton/ToFavoriteButton";
-import SearchButton from "@/components/search/filterSearch/button/SearchButton/SearchButton";
+import SearchButton from "@/components/search/SearchButton/SearchButton";
 import Link from "next/link";
-import WordSearch from "../../search/wordSearch/WordSearch";
+import WordSearch from "../../search/WordSearch/WordSearch";
 import BreadList from "../breadList/BreadList";
 import PageTitle from "../pageTitle/PageTitle";
 import styles from "./Header.module.css";
