+++ src/components/search/FavConditions/FavConditions.tsx
